Add Soulless series, fix gradlew executable bit and finally fix Coalescence and Segmentation!
Soulless series is complete! All one through 6.

S2 is referred to as "Mechanical Machine (Soulless 2)" respectively to its
original title name.

S3 is the all vocals version, referred to as "Soulless 3 (With Vocals)"

S4 is the normal soulless 4 version, not the unnerfed one with the
changed Throwback B audio

S5 is the current released one (not the pre-released audio that Acai
had with the drums during Moonlight Trills)

S6 is... I don't think that's ever changed.

The gradlew executable bit for *nix systems has been changed so now
people cloning this repo can automatically ./gradlew and not have to
chmod +x it.

Coalescence and segmentation I just recoded and redownloaded the audio. It
seems to have worked!